+++ server/src/integrations/currency/types.ts
@@ -9,3 +9,15 @@ export enum CurrencyType{
     Atrium = "atrium",
     Credit = "construct_credit",
 }
+
+export enum MakeMoneyError {
+    NoAccount = "NoAccount",
+    ConnectionError = "ConnectionError",
+    SomethingElse = "SomethingElse",
+    InvalidSign = "InvalidSign",
+    NoFunds = "NoFunds",
+    DuplicateTransaction = "DuplicateTransaction",
+    RefundTooLarge = "RefundTooLarge",
+    Unimplemented = "Unimplemented",
+  };
+
